requirements.txt
# Core Dependencies
discord.py>=2.0.0
PlexAPI>=4.15.0
openai>=1.0.0

# Optional Dependencies
python-dotenv>=1.0.0  # For environment variable management
qbittorrentapi>=2024.5.0  # For torrent management (optional)

# Development Dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0